+++ Scraps.GUI/Extensions/RichTextBoxExtensions.cs
@@ -16,10 +16,6 @@
 /// along with this program. If not, see <https://www.gnu.org/licenses/>.
 #endregion License
 
-using System;
-using System.Drawing;
-using System.Windows.Forms;
-
 namespace Scraps.GUI.Extensions
 {
     public static class RichTextBoxExtensions