+++ Scraps.GUI/Extensions/StringExtensions.cs
@@ -16,8 +16,6 @@
 /// along with this program. If not, see <https://www.gnu.org/licenses/>.
 #endregion License
 
-using System.Text;
-
 namespace Scraps.GUI.Extensions
 {
     public static class StringExtensions