+++ Scraps.GUI/Global.cs
@@ -1,0 +1,9 @@
+﻿global using System.Net;
+global using System.Text;
+global using System.Text.Json;
+global using System.Diagnostics;
+global using System.Text.Json.Serialization;
+
+global using NLog;
+global using NLog.Config;
+global using NLog.Targets;