+++ Scraps.GUI/RaffleRunner/Events/RafflesWonArgs.cs
@@ -16,9 +16,6 @@
 /// along with this program. If not, see <https://www.gnu.org/licenses/>.
 #endregion License
 
-using System;
-using System.Collections.Generic;
-
 namespace Scraps.GUI.RaffleRunner.Events
 {
     public class RafflesWonArgs : EventArgs