+++ Scraps.GUI/RaffleRunner/Events/RunningArgs.cs
@@ -16,8 +16,6 @@
 /// along with this program. If not, see <https://www.gnu.org/licenses/>.
 #endregion License
 
-using System;
-
 namespace Scraps.GUI.RaffleRunner.Events
 {
     public class RunningArgs : EventArgs