+++ Scraps.GUI/Updater/Events/UpdateAvailableArgs.cs
@@ -16,8 +16,6 @@
 /// along with this program. If not, see <https://www.gnu.org/licenses/>.
 #endregion License
 
-using System;
-
 namespace Scraps.GUI.Updater.Events
 {
     public class UpdateAvailableArgs : EventArgs